+++ lib/src/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'modules/dashboard/ui/views/views.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const HomeView(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
