+++ lib/src/modules/auth/auth.dart
@@ -1,0 +1,3 @@
+library com.rick_and_morty.app.auth;
+
+export './ui/routes.dart';
