+++ lib/src/modules/auth/ui/views/login_view.dart
@@ -1,0 +1,15 @@
+part of com.rick_and_morty.app.auth.ui.views;
+
+class LoginView extends StatefulWidget {
+  const LoginView({super.key});
+
+  @override
+  State<LoginView> createState() => _LoginViewState();
+}
+
+class _LoginViewState extends State<LoginView> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
