+++ lib/src/modules/auth/ui/views/views.dart
@@ -1,0 +1,5 @@
+library com.rick_and_morty.app.auth.ui.views;
+
+import 'package:flutter/material.dart';
+
+part './login_view.dart';
