+++ lib/src/modules/dashboard/core/core.dart
@@ -1,0 +1,4 @@
+library com.rick_and_morty.app.dashboard.core;
+
+export './services/services.dart';
+export './state/state.dart';
