+++ lib/src/modules/dashboard/dashboard.dart
@@ -1,0 +1,3 @@
+library com.rick_and_morty.app.dashboard;
+
+export './ui/routes.dart';
