+++ lib/src/modules/dashboard/ui/views/home_view.dart
@@ -1,0 +1,49 @@
+part of com.rick_and_morty.app.dashboard.ui.views;
+
+class HomeView extends StatefulWidget {
+  const HomeView({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<HomeView> createState() => HomeViewState();
+}
+
+class HomeViewState extends State<HomeView> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      _counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'You have pushed the button this many times:',
+            ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
