+++ lib/src/modules/dashboard/ui/views/views.dart
@@ -1,0 +1,5 @@
+library com.rick_and_morty.app.dashboard.ui.views;
+
+import 'package:flutter/material.dart';
+
+part './home_view.dart';
